created array zigzag using cpp.
You are given an array of N. integers. You have to make the given array into a zigzag array. During converting a given array into a zigzag array you can perform a move, which consists of choosing any element and decreasing it by 1.
An array is a zigzag array if either:
1. Every even-indexed element is greater than adjacent elements.
2. OR, every odd-indexed element is greater than adjacent elements. 
Print the minimum number of moves to transform the given array into a zigzag array.